requirements.txt
langchain-openai
langchain
python-dotenv
streamlit
langchain_groq
langchain_community
mysql
mysql-connector-python